finalized core bin_tree and main. currently has opaque, staticly sized borders. this needs to change

+++ bin_tree.h
@@ -48,7 +48,7 @@ struct container
 
 struct rectangle
 {
-	int x, y, width, height;
+	double x, y, width, height;
 };
 
 window *create_window (uint8_t	type, xcb_window_t id);
